+++ packages/utils/index.ts
@@ -1,3 +1,7 @@
 import RafTimer from './RafTimer';
+import sortDataList from './sortDataList';
+import stringToLower from './stringToLower';
+import getElement from './getElement';
+import keyboardMapper from './keyboardMapper';
 
-export { RafTimer };
+export { RafTimer, sortDataList, stringToLower, getElement, keyboardMapper };